requirements.txt
backports-abc==0.5
blinker==1.4
CherryPy==12.0.1
click==6.6
colorama==0.3.9
contextlib2==0.5.4
Cython==0.27.3
decorator==4.1.2
defusedxml==0.4.1
django-allauth==0.28.0
django-material==0.12.3
django-rest-auth==0.9.1
django-rest-framework==0.1.0
django-reversion==2.0.8
Django==1.10.2
djangorestframework==3.5.3
eventlet==0.19.0
Flask-SocketIO==2.7.2
Flask==0.11.1
greenlet==0.4.10
h5py==2.7.0
html5lib==1.0b10
imageio==2.1.2
itsdangerous==0.24
Jinja2==2.10
mxnet==0.12.1
keras-mxnet==1.2.2
livereload==2.5.0
Markdown==2.6.7
MarkupSafe==1.0
networkx==1.11
nvidia-ml-py==7.352.0
oauthlib==2.0.0
opencv-python==3.2.0.7
Pillow==4.3.0
psutil==5.0.0
py-cpuinfo==0.2.3
pycrypto==2.6.1
pydataset==0.2.0
python-dateutil==2.6.1
python-engineio==1.0.3
python-socketio==1.6.0
python3-openid==3.0.10
pytz==2017.2
PyYAML==3.12
raven==6.1.0
redis==2.10.5
requests-oauthlib==0.7.0
scikit-learn==0.18.1
six==1.11.0
sqlparse==0.2.2
tornado==4.5.2
unicodecsv==0.14.1
webencodings==0.5.1
Werkzeug==0.11.11